+++ changelog.md
@@ -8,12 +8,12 @@
 * Infested blocks should generate in every biome 
 * Use Forge to repair items
 * An enchantment to steal air from entities
-* (Quaron) Axes break crops slower
 * Rewrite Harder Crops as a Json feature and make use of block states
   * Due to some blocks having multiple states (e.g. Supplementaries' Flax has bottom and top fully grown)
 
 ## Upcoming
 * Vigour enchantment is now stronger
+* Added a flag config option to wipe the Data Pack list of a world so they are loaded correctly
 * Fixed crops being harder to break with axes
 
 ## 3.5.6